Added a mapping so the popup window information makes more sense.
I don't understand why, but the popup window was listing classes as
fields, methods as variables, functions as values, etc. I forked the
repo and added a little mapping hack that is admittedly neither correct
nor complete. But it fixes my issue for the moment and perhaps helps the
original author find the proper solution.